Add a `copyFragment` helper for plugin onCut/onCopy functions
