+++ docs/reference/slate-react/utils.md
@@ -3,6 +3,7 @@
 
 ```js
 import { 
+  copyFragment,
   findDOMNode,
   findDOMRange,
   findNode,
@@ -17,6 +18,40 @@ React-specific utility functions for Slate that may be useful in certain use cas
 
 
 ## Functions
+
+### `copyFragment`
+`copyFragment(event: DOMEvent|ReactEvent, value: Value, fragment: Document)`
+
+During a cut or copy event, sets `fragment` as the Slate document fragment to be copied.
+
+```js
+function onCopy(event, change, editor) {
+  const { value } = change
+  const fragment = // ... create a fragment from a set of nodes ...
+
+  if (fragment) {
+    copyFragment(event, value, fragment)
+    return true
+  }
+}
+```
+
+Note that calling `copyFragment` should be the last thing you do in your event handler. If you change the window selection after calling `copyFragment`, the browser may copy the wrong content. If you need to perform an action after calling `copyFragment`, wrap it in `requestAnimationFrame`:
+
+```js
+function onCut(event, change, editor) {
+  const { value } = change
+  const fragment = // ... create a fragment from a set of nodes ...
+
+  if (fragment) {
+    copyFragment(event, value, fragment)
+    window.requestAnimationFrame(() => {
+      editor.change(change => change.delete())
+    })
+    return true
+  }
+}
+```
 
 ### `findDOMNode`
 `findDOMNode(node: Node) => DOMElement`
