+++ packages/slate-react/src/index.js
@@ -1,5 +1,6 @@
 
 import Editor from './components/editor'
+import copyFragment from './utils/copy-fragment'
 import findDOMNode from './utils/find-dom-node'
 import findDOMRange from './utils/find-dom-range'
 import findNode from './utils/find-node'
@@ -16,6 +17,7 @@ import setEventTransfer from './utils/set-event-transfer'
 
 export {
   Editor,
+  copyFragment,
   findDOMNode,
   findDOMRange,
   findNode,
@@ -27,6 +29,7 @@ export {
 
 export default {
   Editor,
+  copyFragment,
   findDOMNode,
   findDOMRange,
   findNode,
